Create health insurance project.
The employees of General Manufacturing Corp. receive health insurance through a group plan issued by Wellnet. During the past year, 50 percent of the employees did not file any health insurance claims, 30 percent filed only a small claim, and 20 percent filed a large claim. The small claims were spread uniformly between 0 and $2,000, whereas the large claims were spread uniformly between $2,000 and $20,000. Based on this experience, Wellnet is now negotiating the corporation’s premium payment per employee for the upcoming year and would like to obtain a close estimate of the average cost of insurance coverage for the corporation’s employees.
a) Let U denote the cost to the firm of a small claim, V denote the cost to the firm of a large claim, T denote the type of claim of a randomly chosen employee (no claim, small claim, or large claim), and Y denote the insurance cost to the firm of the employee. Write an expression to relate Y to the other random variables (U, V, and T).
Create an R function entitled “rInsuranceCost” with two parameters m and w. The parameter w represents the breakpoint between small and large claims (currently w = $2,000). The function will return (output) a random vector of size m, where each value of the output vector corresponds to a randomly generated insurance cost from an employee (a value of Y) according to the probability distributions previously described. Report your function by copying-pasting the corresponding code into the Word file.
Write a shiny app for a user interface to input the values of m and w and generate a corresponding random sample of Y. The sample will be generated by calling the rInsuranceCost function. The app will display a histogram and compute a 95% confidence interval for the true average insurance cost to the firm per employee. To input m, use a slider ranging from 100 to 1,000 in increments of 1 (default value is 100). To input w, use a slider ranging from $1,000 to $10,000 in increments of $100 (default value is $2,000). Make sure that your app is user friendly by writing labels to clearly indicate what needs to be input and what the outputs are. Report your shiny app by copying-pasting the corresponding code into the Word file.